requirements-test.txt
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-playwright==0.4.3
requests==2.31.0
faker==20.1.0
